🎉 Task 2完了: PerformanceDataAnalyzer分割 - Main Controller Pattern実装
✅ Issue #95完全達成: PerformanceDataAnalyzer.js分割完了
📊 分割結果:
  • 元ファイル: 2,871語 → メインコントローラー: 876語 (69%削減)
  • 4コンポーネント作成: 全て2,500語以下でMCPツール互換性確保
  • PerformanceMetricsCollector.js: 844語
  • PerformanceDataProcessor.js: 1,433語
  • PerformanceReportGenerator.js: 1,724語
  • PerformanceThresholdManager.js: 1,474語

🏗️ Main Controller Pattern適用:
  • 後方互換性100%維持 - 既存API完全保持
  • 依存注入によるサブコンポーネント統制
  • エラーハンドリング・設定管理統合

次: ErrorHandler.js分割作業
