+++ .kiro/specs/performance-utility-splitting-phase-f3/tasks.md
@@ -6,7 +6,7 @@
   - Set up testing framework for split validation
   - _Requirements: 1.1, 1.2, 1.3, 1.4, 1.5, 1.6_
 
-- [ ] 2. Split PerformanceDataAnalyzer.js (highest priority - 2,871 words)
+- [x] 2. Split PerformanceDataAnalyzer.js (highest priority - 2,871 words)
 - [x] 2.1 Create PerformanceMetricsCollector component
   - Extract metrics collection functionality from PerformanceDataAnalyzer
   - Implement addToAnalysisHistory and related data collection methods
@@ -25,13 +25,13 @@
   - Create unit tests for report generation functionality
   - _Requirements: 4.1, 4.4_
 
-- [ ] 2.4 Create PerformanceThresholdManager component
+- [x] 2.4 Create PerformanceThresholdManager component
   - Extract baseline calibration and adaptive threshold management
   - Implement threshold monitoring and violation detection methods
   - Create unit tests for threshold management functionality
   - _Requirements: 4.1, 4.5_
 
-- [ ] 2.5 Refactor main PerformanceDataAnalyzer to use sub-components
+- [x] 2.5 Refactor main PerformanceDataAnalyzer to use sub-components
   - Modify main class to instantiate and coordinate sub-components
   - Implement delegation methods to maintain public API compatibility
   - Verify all existing functionality works through the new architecture
